+++ internal/topic/topicmultiwriter/interfaces.go
@@ -3,6 +3,7 @@ package topicmultiwriter
 import (
 	"context"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/topic/topicwritercommon"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/topic/topicwriterinternal"
 	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topictypes"
 )
@@ -10,8 +11,7 @@ import (
 type writer interface {
 	Close(ctx context.Context) error
 	WaitInitInfo(ctx context.Context) (topicwriterinternal.InitialInfo, error)
-	Write(ctx context.Context, messages []topicwriterinternal.PublicMessage) error
-	GetBufferedMessages() []topicwriterinternal.PublicMessage
+	WriteInternal(ctx context.Context, messages []topicwritercommon.MessageWithDataContent) error
 }
 
 type writersFactory interface {