+++ internal/topic/topicwriterinternal/writer_transaction.go
@@ -59,7 +59,7 @@ func (w *WriterWithTransaction) WaitInitInfo(ctx context.Context) (InitialInfo, 
 
 func (w *WriterWithTransaction) Write(ctx context.Context, messages []PublicMessage) error {
 	for i := range messages {
-		messages[i].tx = w.tx
+		messages[i].Tx = w.tx
 	}
 
 	return w.streamWriter.Write(ctx, messages)