+++ sql.go
@@ -6,9 +6,12 @@ import (
 	"database/sql/driver"
 	"fmt"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/bind"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/meta"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/secret"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/version"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/xquery"
@@ -46,7 +49,12 @@ func (d *sqlDriver) Open(string) (driver.Conn, error) {
 }
 
 func (d *sqlDriver) OpenConnector(dataSourceName string) (driver.Connector, error) {
-	db, err := Open(context.Background(), dataSourceName)
+	db, err := Open(context.Background(), dataSourceName,
+		With(config.WithBuildInfo(
+			"database/sql",
+			version.Version,
+		)),
+	)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(fmt.Errorf(
 			"failed to connect by data source name '%s': %w",
@@ -259,6 +267,11 @@ type SQLConnector interface {
 }
 
 func Connector(parent *Driver, opts ...ConnectorOption) (SQLConnector, error) {
+	meta.WithBuildInfo(
+		"database/sql",
+		version.Version,
+	)(parent.config.Meta())
+
 	c, err := xsql.Open(parent, parent.metaBalancer, parent.query.Must().Config(),
 		append(
 			append(